+++ typings/custom.d.ts
@@ -1,4 +1,0 @@
-declare module 'lz-string' {
-    var lzString: LZString.LZStringStatic;
-    export = lzString;
-}